Make the server thread a target of itself
