+++ cpp/Osmosis/FileStatus.h
@@ -5,6 +5,7 @@
 #include <sys/stat.h>
 #include <unistd.h>
 #include <boost/algorithm/string/trim.hpp>
+#include <boost/lexical_cast.hpp>
 #include "Common/SplitString.h"
 
 namespace Osmosis
